+++ services/bss/v2/model/model_create_partner_coupons_req.go
@@ -7,23 +7,23 @@ import (
 )
 
 type CreatePartnerCouponsReq struct {
-	// |参数名称：优惠券额度ID优惠券的类型跟随额度中的类型。| |参数约束及描述：优惠券额度ID优惠券的类型跟随额度中的类型。|
+	// 优惠券额度ID。该值在查询优惠券额度接口的响应参数中获取。
 	QuotaId string `json:"quota_id"`
-	// |参数名称：客户ID列表| |参数约束以及描述：客户ID列表|
+	// 客户账号ID。您可以调用查询客户列表接口获取customer_id。
 	CustomerIds []string `json:"customer_ids"`
-	// |参数名称：优惠券的面值：小数点后2位。浮点数精度为：小数点后两位| |参数的约束及描述：优惠券的面值：小数点后2位|
-	FaceValue float32 `json:"face_value"`
-	// |参数名称：优惠券的生效时间,UTC格式：yyyy-MM-ddTHH:mm:ssZ| |参数约束及描述：优惠券的生效时间,UTC格式：yyyy-MM-ddTHH:mm:ssZ|
+	// 代金券面值。 单位：元。取值大于0且精确到小数点后2位。
+	FaceValue float64 `json:"face_value"`
+	// 生效时间。 UTC时间，格式：yyyy-MM-dd'T'HH:mm:ss'Z'，如“2019-05-06T08:05:01Z”。
 	ValidTime *string `json:"valid_time,omitempty"`
-	// |参数名称：优惠券的失效时间,UTC格式：yyyy-MM-ddTHH:mm:ssZ| |参数约束及描述：优惠券的失效时间,UTC格式：yyyy-MM-ddTHH:mm:ssZ|
+	// 失效时间。 UTC时间，格式：yyyy-MM-dd'T'HH:mm:ss'Z'，如“2019-05-06T08:05:01Z”。
 	ExpireTime *string `json:"expire_time,omitempty"`
-	// |参数名称：云服务限制| |参数约束以及描述：云服务限制|
+	// 允许使用的云服务列表，您可以调用查询云服务类型列表接口获取。 多个云服务产品以“,”隔开，最多支持10个。 默认：空（继承额度上的限制） 说明： 如果额度上有限制云服务类型列表，则优惠券上的限制不能超过额度的限制。如果额度上没有限制，则优惠券上可以随意指定云服务类型。
 	CloudServiceTypes *[]string `json:"cloud_service_types,omitempty"`
-	// |参数名称：产品限制| |参数约束以及描述：产品限制|
+	// 允许使用的产品列表。 多个产品以“,”隔开，最多支持10个。 默认：空（继承额度上的限制） 说明： 如果额度上有限制产品列表，则优惠券上的限制不能超过额度的限制。如果额度上没有限制，则优惠券上可以随意指定产品ID。 产品ID需要合作伙伴通过线下获得。
 	ProductIds *[]string `json:"product_ids,omitempty"`
-	// |参数名称：发券时的备注信息| |参数约束及描述：发券时的备注信息|
+	// 发券时的备注信息。
 	Memo *string `json:"memo,omitempty"`
-	// |参数名称：二级经销商ID| |参数约束及描述：如果一级经销商要给二级经销商的子客户设置折扣，需要携带这个字段|
+	// 精英服务商ID。 精英服务商给子客户发放优惠券时，需要携带该参数。
 	IndirectPartnerId *string `json:"indirect_partner_id,omitempty"`
 }
