+++ services/swr/v2/model/model_show_repos_resp.go
@@ -14,13 +14,13 @@ type ShowReposResp struct {
 	// 仓库描述信息
 	Description string `json:"description"`
 	// 仓库大小
-	Size int32 `json:"size"`
+	Size int64 `json:"size"`
 	// 仓库是否为公共仓库，值为true或false
 	IsPublic bool `json:"is_public"`
 	// 仓库中镜像个数，0 ~ 9223372036854775807
-	NumImages int32 `json:"num_images"`
+	NumImages int64 `json:"num_images"`
 	// 仓库下载次数
-	NumDownload int32 `json:"num_download"`
+	NumDownload int64 `json:"num_download"`
 	// 仓库创建时间，UTC日期格式，时间为UTC标准时间，用户需要根据本地时间计算偏移量；如东8区需要+8:00
 	CreatedAt string `json:"created_at"`
 	// 仓库更新时间，UTC日期格式，时间为UTC标准时间，用户需要根据本地时间计算偏移量；如东8区需要+8:00
@@ -42,7 +42,7 @@ type ShowReposResp struct {
 	// 查询他人共享镜像：共享是否过期 查询我共享的镜像：默认为false,无意义
 	Status bool `json:"status"`
 	//
-	TotalRange int32 `json:"total_range"`
+	TotalRange int64 `json:"total_range"`
 }
 
 func (o ShowReposResp) String() string {
